+++ Scripts/Popups/MainPopup/Magnificus/ActMagnificusMapSequence.cs
@@ -21,14 +21,9 @@ public class ActMagnificusMapSequence : BaseMapSequence
     public override void OnGUI()
     {
         bool skipNextNode = ActMagnificus.SkipNextNode;
-        if (Window.Toggle("Skip next node", ref skipNextNode))
+        if (Window.Toggle("Skip map events", ref skipNextNode))
             ToggleSkipNextNode();
 
-        bool activateAllNodes = ActMagnificus.ActivateAllMapNodesActive;
-        if (Window.Toggle("Activate all Map nodes", ref activateAllNodes))
-            ToggleAllNodes();
-
-        Window.Toggle("Toggle Map Override", ref RegionOverride);
         Act.DrawSequencesGUI();
     }
 
@@ -39,15 +34,7 @@ public class ActMagnificusMapSequence : BaseMapSequence
 
     public override void ToggleAllNodes()
     {
-        ActMagnificus.ActivateAllMapNodesActive = !ActMagnificus.ActivateAllMapNodesActive;
-        if (MapNodeManager.m_Instance != null)
-        {
-            MapNode node = Singleton<MapNodeManager>.Instance.ActiveNode;
-            if (node == null)
-                return;
 
-            Singleton<MapNodeManager>.Instance.SetActiveNode(node);
-        }
     }
 
     private Tuple<List<string>, List<string>> RegionNameList()