+++ Plugin.cs
@@ -22,7 +22,7 @@ namespace DebugMenu
     {
         public const string PluginGuid = "jamesgames.inscryption.debugmenu";
         public const string PluginName = "Debug Menu";
-        public const string PluginVersion = "1.5.3";
+        public const string PluginVersion = "1.5.4";
 
         public static Plugin Instance;
         public static ManualLogSource Log;